Remove old transaction history api
