+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/archive/accounts/AccountApiModule.java
@@ -106,15 +106,8 @@ public class AccountApiModule extends AbstractModule {
 
 	@ArchiveEndpoint
 	@ProvidesIntoMap
-	@StringMapKey("account.get_transaction_history2")
+	@StringMapKey("account.get_transaction_history")
 	public JsonRpcHandler accountGetTransactionHistoryReverse(ArchiveAccountHandler archiveAccountHandler) {
 		return archiveAccountHandler::handleAccountGetTransactionHistoryReverse;
-	}
-
-	@ArchiveEndpoint
-	@ProvidesIntoMap
-	@StringMapKey("account.get_transaction_history")
-	public JsonRpcHandler accountGetTransactionHistory(ArchiveAccountHandler archiveAccountHandler) {
-		return archiveAccountHandler::handleAccountGetTransactionHistory;
 	}
 }
