+++ radixdlt-java/radixdlt-java/src/main/java/com/radixdlt/client/lib/api/sync/ImperativeRadixApi.java
@@ -96,7 +96,6 @@ import com.radixdlt.client.lib.dto.TokenBalances;
 import com.radixdlt.client.lib.dto.TokenInfo;
 import com.radixdlt.client.lib.dto.TransactionDTO;
 import com.radixdlt.client.lib.dto.TransactionHistory;
-import com.radixdlt.client.lib.dto.TransactionHistory2;
 import com.radixdlt.client.lib.dto.TransactionStatusDTO;
 import com.radixdlt.client.lib.dto.TxBlobDTO;
 import com.radixdlt.client.lib.dto.TxDTO;
@@ -362,24 +361,12 @@ public interface ImperativeRadixApi {
 
 		/**
 		 * Get transaction history.
-		 * <p>
-		 * To get full list, pass empty cursor for first request and then just pass cursor received in the response
-		 * back to API until you get empty cursor again.
-		 *
-		 * @param address account address for which information is requested
-		 * @param size batch size
-		 * @param cursor pagination cursor
-		 */
-		TransactionHistory history(AccountAddress address, int size, NavigationCursor cursor);
-
-		/**
-		 * Get transaction history.
 		 *
 		 * @param address account address for which information is requested
 		 * @param size batch size
 		 * @param offset offset to start retrieval at
 		 */
-		TransactionHistory2 history2(AccountAddress address, int size, OptionalLong offset);
+		TransactionHistory history(AccountAddress address, int size, OptionalLong offset);
 
 		/**
 		 * Get stakes made from given account.
@@ -665,13 +652,8 @@ public interface ImperativeRadixApi {
 					}
 
 					@Override
-					public TransactionHistory history(AccountAddress address, int size, NavigationCursor cursor) {
-						return unwrap(api.account().history(address, size, Optional.ofNullable(cursor)));
-					}
-
-					@Override
-					public TransactionHistory2 history2(AccountAddress address, int size, OptionalLong offset) {
-						return unwrap(api.account().history2(address, size, offset));
+					public TransactionHistory history(AccountAddress address, int size, OptionalLong offset) {
+						return unwrap(api.account().history(address, size, offset));
 					}
 
 					@Override
