+++ radixdlt-java/radixdlt-java/src/test/java/com/radixdlt/client/lib/api/async/AsyncRadixApiHistoryPaginationTest.java
@@ -67,7 +67,6 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import com.radixdlt.client.lib.api.AccountAddress;
-import com.radixdlt.client.lib.api.NavigationCursor;
 import com.radixdlt.client.lib.api.TransactionRequest;
 import com.radixdlt.client.lib.dto.TransactionDTO;
 import com.radixdlt.client.lib.dto.TransactionHistory;
@@ -78,12 +77,10 @@ import com.radixdlt.utils.Ints;
 import com.radixdlt.utils.UInt256;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.OptionalLong;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 
 /*
@@ -131,16 +128,16 @@ public class AsyncRadixApiHistoryPaginationTest {
 			.onFailure(failure -> fail(failure.toString()))
 			.onSuccess(
 				client -> {
-					var cursorHolder = new AtomicReference<NavigationCursor>();
+					var cursorHolder = new AtomicReference<>(OptionalLong.empty());
 					do {
-						client.account().history(ACCOUNT_ADDRESS1, 50, Optional.ofNullable(cursorHolder.get())).join()
+						client.account().history(ACCOUNT_ADDRESS1, 50, cursorHolder.get()).join()
 							.onFailure(failure -> fail(failure.toString()))
-							.onSuccess(v -> v.getCursor().ifPresent(System.out::println))
-							.onSuccess(v -> v.getCursor().ifPresentOrElse(cursorHolder::set, () -> cursorHolder.set(null)))
+							.onSuccess(v -> v.getNextOffset().ifPresent(System.out::println))
+							.onSuccess(v -> cursorHolder.set(v.getNextOffset()))
 							.map(TransactionHistory::getTransactions)
 							.map(this::formatTxns)
 							.onSuccess(System.out::println);
-					} while (cursorHolder.get() != null && !cursorHolder.get().value().isEmpty());
+					} while (cursorHolder.get().isPresent());
 				});
 	}
 
@@ -172,11 +169,7 @@ public class AsyncRadixApiHistoryPaginationTest {
 		client.transaction().build(request).join()
 			.onFailure(failure -> fail(failure.toString()))
 			.map(builtTransactionDTO -> builtTransactionDTO.toFinalized(KEY_PAIR1))
-			.onSuccess(finalizedTransaction -> client.transaction().finalize(finalizedTransaction, false).join()
-				.onSuccess(txDTO -> assertNotNull(txDTO.getTxId()))
-				.onSuccess(submittableTransaction -> client.transaction().submit(submittableTransaction).join()
-					.onFailure(failure -> fail(failure.toString()))
-					.onSuccess(txDTO -> assertEquals(submittableTransaction.getTxId(), txDTO.getTxId()))));
+			.onSuccess(finalizedTransaction -> client.transaction().finalize(finalizedTransaction, true).join());
 	}
 
 	private static ECKeyPair keyPairOf(int pk) {
