+++ radixdlt-java/radixdlt-java/src/test/java/com/radixdlt/client/lib/api/async/AsyncRadixApiTransactionPaginationTest.java
@@ -66,7 +66,7 @@ package com.radixdlt.client.lib.api.async;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import com.radixdlt.client.lib.dto.Transaction2DTO;
+import com.radixdlt.client.lib.dto.TransactionDTO;
 import com.radixdlt.client.lib.dto.TransactionsDTO;
 
 import java.util.List;
@@ -110,7 +110,7 @@ public class AsyncRadixApiTransactionPaginationTest {
 				});
 	}
 
-	private List<String> formatTxns(List<Transaction2DTO> t) {
+	private List<String> formatTxns(List<TransactionDTO> t) {
 		return t.stream()
 			.map(v -> String.format(
 				"%s (%s) - %s (%d:%d), Fee: %s%n",
