+++ radixdlt-java/radixdlt-java/src/test/java/com/radixdlt/client/lib/dto/DtoTest.java
@@ -372,17 +372,12 @@ public class DtoTest {
 
 	@Test
 	public void testTransaction2DTO() {
-		EqualsVerifier.forClass(Transaction2DTO.class).suppress(Warning.NULL_FIELDS).verify();
-	}
-
-	@Test
-	public void testTransactionHistory() {
-		EqualsVerifier.forClass(TransactionHistory.class).suppress(Warning.NULL_FIELDS).verify();
+		EqualsVerifier.forClass(TransactionDTO.class).suppress(Warning.NULL_FIELDS).verify();
 	}
 
 	@Test
 	public void testTransactionHistory2() {
-		EqualsVerifier.forClass(TransactionHistory2.class).suppress(Warning.NULL_FIELDS).verify();
+		EqualsVerifier.forClass(TransactionHistory.class).suppress(Warning.NULL_FIELDS).verify();
 	}
 
 	@Test
